Frontend successfully receives data from datastore.

+++ portfolio/src/main/java/com/google/sps/data/Service.java
@@ -7,7 +7,7 @@ public final class Service {
   private final String service_name;
   private final String service_description;
   private final long provider_id;
-  private final long average_rating;
+  private final Double average_rating;
 
   /**
    * @param {!long} id The unique identifier for each Service.
@@ -16,7 +16,7 @@ public final class Service {
    * @param {!long} provider The id of this Service's provider.
    * @param {!long} averageRating This Service's average Rating.
    */
-  public Service(long service_id, String service_name, String service_description, long provider_id, long average_rating) {
+  public Service(long service_id, String service_name, String service_description, long provider_id, Double average_rating) {
       this.service_id = service_id;
       this.service_name = service_name;
       this.service_description = service_description;
@@ -45,7 +45,7 @@ public final class Service {
   }
 
   /** Getter method for average rating. */
-  public long getAverage() {
+  public Double getAverage() {
     return average_rating;
   }
 }
